refactor: field validator into an interface
