+++ core/src/main/java/me/naingaungluu/formconductor/validation/validators/FieldValidator.kt
@@ -2,6 +2,6 @@ package me.naingaungluu.formconductor.validation.validators
 
 import me.naingaungluu.formconductor.FieldResult
 
-internal abstract class FieldValidator<V: Any?> {
-    abstract fun validate(input: V): FieldResult
+internal interface FieldValidator<V: Any?> {
+    fun validate(input: V): FieldResult
 }
